handle utf-16 surrogate pairs

+++ inc/utf8.h
@@ -19,10 +19,11 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #ifndef MONITOR_UTF8_H
 #define MONITOR_UTF8_H
 
+#include <stdint.h>
 #include <windows.h>
 
-int utf8_encode(unsigned short x, unsigned char *out);
-int utf8_length(unsigned short x);
+int utf8_encode(uint32_t x, unsigned char *out);
+int utf8_length(uint32_t x);
 
 int utf8_bytecnt_ascii(const char *s, int len);
 int utf8_bytecnt_unicode(const wchar_t *s, int len);
